refactor: Refactors drunk level to use doubles for precision
Modifies the 'drunkLevel' field in the DrinkingRecord model and related services to use double data type.
This change enables more precise representation and calculation of average drunk levels,
improving the accuracy of displayed data in the calendar, profile stats, and reports.

The SakuCharacter widget and drunk level text displays are also updated to display integers derived from the drunk level,
ensuring a more user-friendly interface, while retaining the improved precision in the backend calculations.

+++ lib/features/calendar/widgets/dialogs/add_record_dialog.dart
@@ -187,7 +187,7 @@ class _AddRecordDialogState extends ConsumerState<AddRecordDialog> {
         date: widget.selectedDate,
         sessionNumber: 0, // 서비스에서 자동 계산
         meetingName: _meetingNameController.text,
-        drunkLevel: _drunkLevel.toInt(),
+        drunkLevel: _drunkLevel,
         yearMonth: DateFormat(
           'yyyy-MM',
         ).format(widget.selectedDate), // Added yearMonth
